requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1
django-allauth==64.1.0
django-environ==0.11.2
django-stubs==5.0.4
django-stubs-ext==5.0.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djangorestframework-stubs==3.15.1
drf-yasg==1.21.7
idna==3.10
inflection==0.5.1
packaging==24.1
psycopg==3.2.1
PyJWT==2.9.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.1
types-PyYAML==6.0.12.20240808
types-requests==2.32.0.20240914
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
